+++ extensions/pastery/src/types.ts
@@ -1,13 +1,14 @@
 export type Paste = {
-    duration: number;
+    duration: number | null;
     id: string;
     language: string;
     title: string;
     url: string;
 }
-export type ActionResult = {
-    result: "success"
-} | {
+export type ErrorResult = {
     result: "error";
     error_msg: string;
-}
+};
+export type ActionResult = {
+    result: "success"
+} | ErrorResult;