+++ extensions/pastery/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Pastery Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Search Pastes
+- Create Paste
+- Delete Paste